# apps/bilgeapi/docs/openapi/bilgeapi_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BilgeAPI",
    "description": "Independent Incident Intake, Diagnostic and Repair-Orchestration API",
    "version": "1.3.0"
  },
  "paths": {
...
    "/v1/incidents": {
      "post": {
        "tags": [
          "Incidents"
        ],
        "summary": "Create Incident",
        "operationId": "create_incident_v1_incidents_post",
        "security": [
          {
            "ApiKeyHeader": []
          },
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Incidents"
        ],
        "summary": "List Incidents",
        "operationId": "list_incidents_v1_incidents_get",
        "security": [
          {
            "ApiKeyHeader": []
          },
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "project_key",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter incidents by project key",
              "title": "Project Key"
            },
            "description": "Filter incidents by project key"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IncidentResponse"
                  },
                  "title": "Response List Incidents V1 Incidents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/incidents/{incident_id}": {
      "get": {
        "tags": [
          "Incidents"
        ],
        "summary": "Get Incident",
        "operationId": "get_incident_v1_incidents__incident_id__get",
        "security": [
          {
            "ApiKeyHeader": []
          },
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/incidents/{incident_id}/diagnostics": {
      "post": {
        "tags": [
          "Diagnostics"
        ],
        "summary": "Trigger Diagnostic",
        "operationId": "trigger_diagnostic_v1_incidents__incident_id__diagnostics_post",
        "security": [
          {
            "ApiKeyHeader": []
          },
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/diagnostics": {
      "get": {
        "tags": [
          "Diagnostics"
        ],
        "summary": "List Diagnostics",
        "operationId": "list_diagnostics_v1_diagnostics_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DiagnosticResult"
                  },
                  "type": "array",
                  "title": "Response List Diagnostics V1 Diagnostics Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyHeader": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/diagnostics/{diagnostic_id}": {
      "get": {
        "tags": [
          "Diagnostics"
        ],
        "summary": "Get Diagnostic",
        "operationId": "get_diagnostic_v1_diagnostics__diagnostic_id__get",
        "security": [
          {
            "ApiKeyHeader": []
          },
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "diagnostic_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Diagnostic Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DiagnosticResult": {
        "properties": {
          "diagnostic_id": {
            "type": "string",
            "title": "Diagnostic Id"
          },
          "incident_id": {
            "type": "string",
            "title": "Incident Id"
          },
          "status": {
            "$ref": "#/components/schemas/DiagnosticStatus"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "root_cause_hypothesis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Root Cause Hypothesis"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "risk_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Score"
          },
          "findings": {
            "items": {},
            "type": "array",
            "title": "Findings"
          },
          "recommendations": {
            "items": {},
            "type": "array",
            "title": "Recommendations"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "diagnostic_id",
          "incident_id",
          "status",
          "created_at"
        ],
        "title": "DiagnosticResult"
      },
      "DiagnosticStartResponse": {
        "properties": {
          "diagnostic_id": {
            "type": "string",
            "title": "Diagnostic Id"
          },
          "status": {
            "$ref": "#/components/schemas/DiagnosticStatus",
            "default": "QUEUED"
          },
          "next_action": {
            "type": "string",
            "title": "Next Action",
            "default": "CHECK_DIAGNOSTIC_STATUS"
          }
        },
        "type": "object",
        "required": [
          "diagnostic_id"
        ],
        "title": "DiagnosticStartResponse"
      },
      "DiagnosticStatus": {
        "type": "string",
        "enum": [
          "QUEUED",
          "RUNNING",
          "COMPLETED",
          "FAILED"
        ],
        "title": "DiagnosticStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IncidentCreate": {
        "properties": {
          "project_key": {
            "type": "string",
            "title": "Project Key"
          },
          "source_system": {
            "type": "string",
            "title": "Source System"
          },
          "environment": {
            "type": "string",
            "title": "Environment"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "error_message": {
            "type": "string",
            "minLength": 1,
            "title": "Error Message"
          },
          "stack_trace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stack Trace"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "project_key",
          "source_system",
          "environment",
          "kind",
          "severity",
          "error_message",
          "occurred_at"
        ],
        "title": "IncidentCreate"
      },
      "IncidentResponse": {
        "properties": {
          "project_key": {
            "type": "string",
            "title": "Project Key"
          },
          "source_system": {
            "type": "string",
            "title": "Source System"
          },
          "environment": {
            "type": "string",
            "title": "Environment"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "error_message": {
            "type": "string",
            "minLength": 1,
            "title": "Error Message"
          },
          "stack_trace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stack Trace"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "project_key",
          "source_system",
          "environment",
          "kind",
          "severity",
          "error_message",
          "occurred_at",
          "id",
          "created_at"
        ],
        "title": "IncidentResponse"
      },
...
      "Severity": {
        "type": "string",
        "enum": [
          "CRITICAL",
          "HIGH",
          "MEDIUM",
          "LOW"
        ],
        "title": "Severity"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "ApiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}